Chnaged the rejectPayment() function to properly handle the students payment rejection who want to upgrade. There was a huge bug which was doing even their previous payment's rejection too. Moreover, changed the getUnverifiedUsers() function to send only those users who have uploaded their payment receipt.

+++ backend/controllers/payment.js
@@ -2,13 +2,25 @@ import { AppError, handleAsyncError } from "../error.js";
 import jwt from "jsonwebtoken";
 import pool from "../database.js";
 import { unlink } from "node:fs/promises";
+import fs from "fs/promises";
 
 export const isCouponValid = async (coupon) => {
     return (await pool.query("SELECT code FROM coupons WHERE code=$1", [coupon])).rows[0] ? true : false;
 };
 
 export const getUnverifiedUsers = handleAsyncError(async (req, res, next) => {
-    const data = (await pool.query("SELECT name, email, phone, role, upgrade_role, academic_status, payment_status, upgrade_status, coupon FROM users INNER JOIN students ON students.student_id=users.user_id WHERE payment_status='PENDING' OR upgrade_status='PENDING' ORDER BY user_id ASC")).rows;
+    let emails = [];
+
+    try {
+        emails = await fs.readdir("receipts");
+    } catch (err) {
+        console.error("Error: Receipts directory doesn't exist!", err);
+    }
+    
+    emails = emails.map(email => email.replace(".jpg", ""));
+
+    const data = (await pool.query("SELECT name, email, phone, role, upgrade_role, academic_status, payment_status, upgrade_status, coupon FROM users INNER JOIN students ON students.student_id=users.user_id WHERE email = ANY($1) ORDER BY user_id ASC", [emails])).rows;
+
     res.status(200).json({
         status: "success",
         data
@@ -17,8 +29,19 @@ export const getUnverifiedUsers = handleAsyncError(async (req, res, next) => {
 export const rejectPayment = handleAsyncError(async (req, res, next) => {
     if (!req.body?.email)
         return next(new AppError("Please provide the user's email to verify his/her payment", 400));
-    await pool.query("UPDATE students SET payment_status='REJECTED' WHERE student_id = (SELECT user_id FROM users WHERE email=$1)", [req.body.email]);
-    await unlink(`receipts/${req.body.email}.jpg`);
+    if (!req.body?.payment_status)
+        return next(new AppError("Please provide user's current payment status to reject his/her payment", 400));
+
+    if (req.body.payment_status === 'VERIFIED') {
+        await pool.query("UPDATE students SET upgrade_status=NULL, upgrade_role=NULL WHERE student_id = (SELECT user_id FROM users WHERE email=$1)", [req.body.email]);
+    } else {
+        await pool.query("UPDATE students SET payment_status='REJECTED', upgrade_status=NULL, upgrade_role=NULL WHERE student_id = (SELECT user_id FROM users WHERE email=$1)", [req.body.email]);
+    }
+    try {
+        await unlink(`receipts/${req.body.email}.jpg`);
+    } catch(err) {
+        console.error(`Error deleting receipt for ${req.body.email}:`, err);
+    }
     res.status(200).json({
         status: "success",
         message: "Payment Rejected"
@@ -33,8 +56,11 @@ export const verifyPayment = handleAsyncError(async (req, res, next) => {
 
     await pool.query("UPDATE students SET payment_status='VERIFIED', upgrade_status=NULL WHERE student_id = (SELECT user_id FROM users WHERE email=$1)", [req.body.email]);
     await pool.query("UPDATE users SET role=$2 WHERE email=$1", [req.body.email, req.body.role]);
-    await unlink(`receipts/${req.body.email}.jpg`);
-
+    try {
+        await unlink(`receipts/${req.body.email}.jpg`);
+    } catch(err) {
+        console.error(`Error deleting receipt for ${req.body.email}:`, err);
+    }
     res.status(200).json({
         status: "success",
         message: "Payment Verified"
